+++ Packages/com.unity.ide.vscode/Editor/ProjectGeneration/ProjectGeneration.cs
@@ -5,10 +5,8 @@ using System.Linq;
 using System.Security;
 using System.Security.Cryptography;
 using System.Text;
-using System.Text.RegularExpressions;
 using UnityEditor;
 using UnityEditor.Compilation;
-using UnityEditor.PackageManager;
 using UnityEngine;
 using UnityEngine.Profiling;
 
@@ -20,6 +18,7 @@ namespace VSCodeEditor
         void Sync();
         string SolutionFile();
         string ProjectDirectory { get; }
+        IAssemblyNameProvider AssemblyNameProvider { get; }
         void GenerateAll(bool generateAll);
         bool SolutionExists();
     }
@@ -112,17 +111,27 @@ namespace VSCodeEditor
 
         string m_SolutionProjectEntryTemplate = string.Join("\r\n", @"Project(""{{{0}}}"") = ""{1}"", ""{2}"", ""{{{3}}}""", @"EndProject").Replace("    ", "\t");
 
-        string m_SolutionProjectConfigurationTemplate = string.Join("\r\n", @"        {{{0}}}.Debug|Any CPU.ActiveCfg = Debug|Any CPU", @"        {{{0}}}.Debug|Any CPU.Build.0 = Debug|Any CPU", @"        {{{0}}}.Release|Any CPU.ActiveCfg = Release|Any CPU", @"        {{{0}}}.Release|Any CPU.Build.0 = Release|Any CPU").Replace("    ", "\t");
+        string m_SolutionProjectConfigurationTemplate = string.Join("\r\n", @"        {{{0}}}.Debug|Any CPU.ActiveCfg = Debug|Any CPU", @"        {{{0}}}.Debug|Any CPU.Build.0 = Debug|Any CPU").Replace("    ", "\t");
 
         static readonly string[] k_ReimportSyncExtensions = { ".dll", ".asmdef" };
 
         string[] m_ProjectSupportedExtensions = new string[0];
         public string ProjectDirectory { get; }
-        bool m_ShouldGenerateAll;
+        IAssemblyNameProvider IGenerator.AssemblyNameProvider => m_AssemblyNameProvider;
 
         public void GenerateAll(bool generateAll)
         {
-            m_ShouldGenerateAll = generateAll;
+            m_AssemblyNameProvider.ToggleProjectGeneration(
+                ProjectGenerationFlag.BuiltIn
+                | ProjectGenerationFlag.Embedded
+                | ProjectGenerationFlag.Git
+                | ProjectGenerationFlag.Local
+#if UNITY_2019_3_OR_NEWER
+                | ProjectGenerationFlag.LocalTarBall
+#endif
+                | ProjectGenerationFlag.PlayerAssemblies
+                | ProjectGenerationFlag.Registry
+                | ProjectGenerationFlag.Unknown);
         }
 
         readonly string m_ProjectName;
@@ -172,8 +181,8 @@ namespace VSCodeEditor
                 var allProjectAssemblies = RelevantAssembliesForMode(assemblies).ToList();
                 var allAssetProjectParts = GenerateAllAssetProjectParts();
 
-                var affectedNames = affectedFiles.Select(asset => m_AssemblyNameProvider.GetAssemblyNameFromScriptPath(asset)?.Split(new [] {".dll"}, StringSplitOptions.RemoveEmptyEntries)[0]);
-                var reimportedNames = reimportedFiles.Select(asset => m_AssemblyNameProvider.GetAssemblyNameFromScriptPath(asset)?.Split(new [] {".dll"}, StringSplitOptions.RemoveEmptyEntries)[0]);
+                var affectedNames = affectedFiles.Select(asset => m_AssemblyNameProvider.GetAssemblyNameFromScriptPath(asset)).Where(name => !string.IsNullOrWhiteSpace(name)).Select(name => name.Split(new [] {".dll"}, StringSplitOptions.RemoveEmptyEntries)[0]);
+                var reimportedNames = reimportedFiles.Select(asset => m_AssemblyNameProvider.GetAssemblyNameFromScriptPath(asset)).Where(name => !string.IsNullOrWhiteSpace(name)).Select(name => name.Split(new [] {".dll"}, StringSplitOptions.RemoveEmptyEntries)[0]);
                 var affectedAndReimported = new HashSet<string>(affectedNames.Concat(reimportedNames));
                 var assemblyNames = new HashSet<string>(allProjectAssemblies.Select(assembly => Path.GetFileName(assembly.outputPath)));
 
@@ -216,7 +225,7 @@ namespace VSCodeEditor
 
         void SetupProjectSupportedExtensions()
         {
-            m_ProjectSupportedExtensions = EditorSettings.projectGenerationUserExtensions;
+            m_ProjectSupportedExtensions = m_AssemblyNameProvider.ProjectSupportedExtensions;
         }
 
         bool ShouldFileBePartOfSolution(string file)
@@ -224,7 +233,7 @@ namespace VSCodeEditor
             string extension = Path.GetExtension(file);
 
             // Exclude files coming from packages except if they are internalized.
-            if (!m_ShouldGenerateAll && IsInternalizedPackagePath(file))
+            if (m_AssemblyNameProvider.IsInternalizedPackagePath(file))
             {
                 return false;
             }
@@ -293,7 +302,7 @@ namespace VSCodeEditor
             WriteVSCodeSettingsFiles();
         }
 
-        IEnumerable<ResponseFileData> ParseResponseFileData(Assembly assembly)
+        List<ResponseFileData> ParseResponseFileData(Assembly assembly)
         {
             var systemReferenceDirectories = CompilationPipeline.GetSystemAssemblyDirectories(assembly.compilerOptions.ApiCompatibilityLevel);
 
@@ -315,7 +324,7 @@ namespace VSCodeEditor
                 }
             }
 
-            return responseFilesData.Select(x => x.Value);
+            return responseFilesData.Select(x => x.Value).ToList();
         }
 
         Dictionary<string, string> GenerateAllAssetProjectParts()
@@ -326,7 +335,7 @@ namespace VSCodeEditor
             {
                 // Exclude files coming from packages except if they are internalized.
                 // TODO: We need assets from the assembly API
-                if (!m_ShouldGenerateAll && IsInternalizedPackagePath(asset))
+                if (m_AssemblyNameProvider.IsInternalizedPackagePath(asset))
                 {
                     continue;
                 }
@@ -362,30 +371,18 @@ namespace VSCodeEditor
             return result;
         }
 
-        bool IsInternalizedPackagePath(string file)
-        {
-            if (string.IsNullOrWhiteSpace(file))
-            {
-                return false;
-            }
-
-            var packageInfo = m_AssemblyNameProvider.FindForAssetPath(file);
-            if (packageInfo == null)
-            {
-                return false;
-            }
-
-            var packageSource = packageInfo.source;
-            return packageSource != PackageSource.Embedded && packageSource != PackageSource.Local;
-        }
-
         void SyncProject(
             Assembly assembly,
             Dictionary<string, string> allAssetsProjectParts,
-            IEnumerable<ResponseFileData> responseFilesData,
+            List<ResponseFileData> responseFilesData,
             HashSet<string> assemblyNames)
         {
-            SyncProjectFileIfNotChanged(ProjectFile(assembly), ProjectText(assembly, allAssetsProjectParts, responseFilesData, assemblyNames));
+            SyncProjectFileIfNotChanged(ProjectFile(assembly), ProjectText(assembly, allAssetsProjectParts, responseFilesData, assemblyNames, GetAllRoslynAnalyzerPaths().ToArray()));
+        }
+
+        private IEnumerable<string> GetAllRoslynAnalyzerPaths()
+        {
+            return m_AssemblyNameProvider.GetRoslynAnalyzerPaths();
         }
 
         void SyncProjectFileIfNotChanged(string path, string newContents)
@@ -416,13 +413,13 @@ namespace VSCodeEditor
         string ProjectText(
             Assembly assembly,
             Dictionary<string, string> allAssetsProjectParts,
-            IEnumerable<ResponseFileData> responseFilesData,
-            HashSet<string> assemblyNames)
+            List<ResponseFileData> responseFilesData,
+            HashSet<string> assemblyNames,
+            string[] roslynAnalyzerDllPaths)
         {
             var projectBuilder = new StringBuilder();
-            ProjectHeader(assembly, responseFilesData, projectBuilder);
+            ProjectHeader(assembly, responseFilesData, roslynAnalyzerDllPaths, projectBuilder);
             var references = new List<string>();
-            var projectReferences = new List<Assembly>();
 
             foreach (string file in assembly.sourceFiles)
             {
@@ -446,7 +443,16 @@ namespace VSCodeEditor
                 projectBuilder.Append(additionalAssetsForProject);
 
             var responseRefs = responseFilesData.SelectMany(x => x.FullPathReferences.Select(r => r));
-            foreach (var reference in assembly.compiledAssemblyReferences.Union(responseRefs).Union(references))
+            var internalAssemblyReferences = assembly.assemblyReferences
+              .Where(i => !i.sourceFiles.Any(ShouldFileBePartOfSolution)).Select(i => i.outputPath);
+            var allReferences =
+              assembly.compiledAssemblyReferences
+                .Union(responseRefs)
+                .Union(references)
+                .Union(internalAssemblyReferences)
+                .Except(roslynAnalyzerDllPaths);
+
+            foreach (var reference in allReferences)
             {
                 string fullReference = Path.IsPathRooted(reference) ? reference : Path.Combine(ProjectDirectory, reference);
                 AppendReference(fullReference, projectBuilder);
@@ -456,13 +462,14 @@ namespace VSCodeEditor
             {
                 projectBuilder.Append("  </ItemGroup>").Append(k_WindowsNewline);
                 projectBuilder.Append("  <ItemGroup>").Append(k_WindowsNewline);
-                foreach (Assembly reference in assembly.assemblyReferences)
+                foreach (Assembly reference in assembly.assemblyReferences.Where(i => i.sourceFiles.Any(ShouldFileBePartOfSolution)))
                 {
                     var referencedProject = reference.outputPath;
 
                     projectBuilder.Append("    <ProjectReference Include=\"").Append(reference.name).Append(GetProjectExtension()).Append("\">").Append(k_WindowsNewline);
                     projectBuilder.Append("      <Project>{").Append(ProjectGuid(reference.name)).Append("}</Project>").Append(k_WindowsNewline);
                     projectBuilder.Append("      <Name>").Append(reference.name).Append("</Name>").Append(k_WindowsNewline);
+                    projectBuilder.Append("      <ReferenceOutputAssembly>false</ReferenceOutputAssembly>").Append(k_WindowsNewline);
                     projectBuilder.Append("    </ProjectReference>").Append(k_WindowsNewline);
                 }
             }
@@ -477,9 +484,9 @@ namespace VSCodeEditor
             var escapedFullPath = SecurityElement.Escape(fullReference);
             escapedFullPath = escapedFullPath.Replace("\\\\", "/");
             escapedFullPath = escapedFullPath.Replace("\\", "/");
-            projectBuilder.Append(" <Reference Include=\"").Append(Path.GetFileNameWithoutExtension(escapedFullPath)).Append("\">").Append(k_WindowsNewline);
-            projectBuilder.Append(" <HintPath>").Append(escapedFullPath).Append("</HintPath>").Append(k_WindowsNewline);
-            projectBuilder.Append(" </Reference>").Append(k_WindowsNewline);
+            projectBuilder.Append("    <Reference Include=\"").Append(Path.GetFileNameWithoutExtension(escapedFullPath)).Append("\">").Append(k_WindowsNewline);
+            projectBuilder.Append("        <HintPath>").Append(escapedFullPath).Append("</HintPath>").Append(k_WindowsNewline);
+            projectBuilder.Append("    </Reference>").Append(k_WindowsNewline);
         }
 
         public string ProjectFile(Assembly assembly)
@@ -496,22 +503,75 @@ namespace VSCodeEditor
 
         void ProjectHeader(
             Assembly assembly,
-            IEnumerable<ResponseFileData> responseFilesData,
+            List<ResponseFileData> responseFilesData,
+            string[] roslynAnalyzerDllPaths,
             StringBuilder builder
         )
         {
-            // TODO: .Concat(EditorUserBuildSettings.activeScriptCompilationDefines)
+            var otherArguments = GetOtherArgumentsFromResponseFilesData(responseFilesData);
             GetProjectHeaderTemplate(
                 builder,
                 ProjectGuid(assembly.name),
                 assembly.name,
-                string.Join(";", new[] { "DEBUG", "TRACE" }.Concat(assembly.defines).Concat(responseFilesData.SelectMany(x => x.Defines)).Distinct().ToArray()),
-                assembly.compilerOptions.AllowUnsafeCode | responseFilesData.Any(x => x.Unsafe));
+                string.Join(";", new[] { "DEBUG", "TRACE" }.Concat(assembly.defines).Concat(responseFilesData.SelectMany(x => x.Defines)).Concat(EditorUserBuildSettings.activeScriptCompilationDefines).Distinct().ToArray()),
+                assembly.compilerOptions.AllowUnsafeCode | responseFilesData.Any(x => x.Unsafe),
+                GenerateAnalyserItemGroup(otherArguments["analyzer"].Concat(otherArguments["a"])
+                    .SelectMany(x => x.Split(';'))
+                    .Concat(roslynAnalyzerDllPaths)
+                    .Distinct()
+                    .ToArray()));
+        }
+
+        private static ILookup<string, string> GetOtherArgumentsFromResponseFilesData(List<ResponseFileData> responseFilesData)
+        {
+            var paths = responseFilesData.SelectMany(x =>
+                {
+                    return x.OtherArguments.Where(a => a.StartsWith("/") || a.StartsWith("-"))
+                                           .Select(b =>
+                    {
+                        var index = b.IndexOf(":", StringComparison.Ordinal);
+                        if (index > 0 && b.Length > index)
+                        {
+                            var key = b.Substring(1, index - 1);
+                            return new KeyValuePair<string, string>(key, b.Substring(index + 1));
+                        }
+
+                        const string warnaserror = "warnaserror";
+                        if (b.Substring(1).StartsWith(warnaserror))
+                        {
+                            return new KeyValuePair<string, string>(warnaserror, b.Substring(warnaserror.Length + 1));
+                        }
+
+                        return default;
+                    });
+                })
+              .Distinct()
+              .ToLookup(o => o.Key, pair => pair.Value);
+            return paths;
+        }
+
+        private static string GenerateAnalyserItemGroup(string[] paths)
+        {
+            //    <ItemGroup>
+            //        <Analyzer Include="..\packages\Comments_analyser.1.0.6626.21356\analyzers\dotnet\cs\Comments_analyser.dll" />
+            //        <Analyzer Include="..\packages\UnityEngineAnalyzer.1.0.0.0\analyzers\dotnet\cs\UnityEngineAnalyzer.dll" />
+            //    </ItemGroup>
+            if (!paths.Any())
+                return string.Empty;
+
+            var analyserBuilder = new StringBuilder();
+            analyserBuilder.Append("  <ItemGroup>").Append(k_WindowsNewline);
+            foreach (var path in paths)
+            {
+                analyserBuilder.Append($"    <Analyzer Include=\"{path}\" />").Append(k_WindowsNewline);
+            }
+            analyserBuilder.Append("  </ItemGroup>").Append(k_WindowsNewline);
+            return analyserBuilder.ToString();
         }
 
         static string GetSolutionText()
         {
-            return string.Join("\r\n", @"", @"Microsoft Visual Studio Solution File, Format Version {0}", @"# Visual Studio {1}", @"{2}", @"Global", @"    GlobalSection(SolutionConfigurationPlatforms) = preSolution", @"        Debug|Any CPU = Debug|Any CPU", @"        Release|Any CPU = Release|Any CPU", @"    EndGlobalSection", @"    GlobalSection(ProjectConfigurationPlatforms) = postSolution", @"{3}", @"    EndGlobalSection", @"    GlobalSection(SolutionProperties) = preSolution", @"        HideSolutionNode = FALSE", @"    EndGlobalSection", @"EndGlobal", @"").Replace("    ", "\t");
+            return string.Join("\r\n", @"", @"Microsoft Visual Studio Solution File, Format Version {0}", @"# Visual Studio {1}", @"{2}", @"Global", @"    GlobalSection(SolutionConfigurationPlatforms) = preSolution", @"        Debug|Any CPU = Debug|Any CPU", @"    EndGlobalSection", @"    GlobalSection(ProjectConfigurationPlatforms) = postSolution", @"{3}", @"    EndGlobalSection", @"    GlobalSection(SolutionProperties) = preSolution", @"        HideSolutionNode = FALSE", @"    EndGlobalSection", @"EndGlobal", @"").Replace("    ", "\t");
         }
 
         static string GetProjectFooterTemplate()
@@ -524,7 +584,8 @@ namespace VSCodeEditor
             string assemblyGUID,
             string assemblyName,
             string defines,
-            bool allowUnsafe
+            bool allowUnsafe,
+            string analyzerBlock
         )
         {
             builder.Append(@"<?xml version=""1.0"" encoding=""utf-8""?>").Append(k_WindowsNewline);
@@ -557,15 +618,6 @@ namespace VSCodeEditor
             builder.Append(@"    <NoWarn>0169</NoWarn>").Append(k_WindowsNewline);
             builder.Append(@"    <AllowUnsafeBlocks>").Append(allowUnsafe).Append("</AllowUnsafeBlocks>").Append(k_WindowsNewline);
             builder.Append(@"  </PropertyGroup>").Append(k_WindowsNewline);
-            builder.Append(@"  <PropertyGroup Condition="" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' "">").Append(k_WindowsNewline);
-            builder.Append(@"    <DebugType>pdbonly</DebugType>").Append(k_WindowsNewline);
-            builder.Append(@"    <Optimize>true</Optimize>").Append(k_WindowsNewline);
-            builder.Append(@"    <OutputPath>Temp\bin\Release\</OutputPath>").Append(k_WindowsNewline);
-            builder.Append(@"    <ErrorReport>prompt</ErrorReport>").Append(k_WindowsNewline);
-            builder.Append(@"    <WarningLevel>4</WarningLevel>").Append(k_WindowsNewline);
-            builder.Append(@"    <NoWarn>0169</NoWarn>").Append(k_WindowsNewline);
-            builder.Append(@"    <AllowUnsafeBlocks>").Append(allowUnsafe).Append("</AllowUnsafeBlocks>").Append(k_WindowsNewline);
-            builder.Append(@"  </PropertyGroup>").Append(k_WindowsNewline);
             builder.Append(@"  <PropertyGroup>").Append(k_WindowsNewline);
             builder.Append(@"    <NoConfig>true</NoConfig>").Append(k_WindowsNewline);
             builder.Append(@"    <NoStdLib>true</NoStdLib>").Append(k_WindowsNewline);
@@ -573,6 +625,7 @@ namespace VSCodeEditor
             builder.Append(@"    <ImplicitlyExpandNETStandardFacades>false</ImplicitlyExpandNETStandardFacades>").Append(k_WindowsNewline);
             builder.Append(@"    <ImplicitlyExpandDesignTimeFacades>false</ImplicitlyExpandDesignTimeFacades>").Append(k_WindowsNewline);
             builder.Append(@"  </PropertyGroup>").Append(k_WindowsNewline);
+            builder.Append(analyzerBlock);
             builder.Append(@"  <ItemGroup>").Append(k_WindowsNewline);
         }